Combine trim_parameters.py and trimming.py
