+++ tests/virtool_workflow/analysis/test_skewer.py
@@ -9,7 +9,7 @@ from tests.virtool_workflow.api.mocks.mock_sample_routes import (
 )
 from virtool_workflow import features, Workflow
 from virtool_workflow.abc.caches.analysis_caches import ReadsCache
-from virtool_workflow.analysis.features.trimming import Trimming
+from virtool_workflow.analysis.trimming import Trimming
 from virtool_workflow.analysis.read_prep.skewer import skewer, trimming_min_length
 from virtool_workflow.caching.local import LocalCaches
 from virtool_workflow.data_model import Job
