+++ virtool_workflow/analysis/trimming.py
@@ -1,12 +1,17 @@
+"""Calculate trimming parameters which are passed the Skewer read trimming tool."""
 import hashlib
 import json
 import shutil
 from functools import partial
 from pathlib import Path
+from typing import Dict, Union
 
 import aiohttp
+from virtool_core.samples.utils import TRIM_PARAMETERS
+
 from virtool_workflow import FixtureScope
 from virtool_workflow.abc.caches.analysis_caches import ReadsCache
+from virtool_workflow.analysis.library_types import LibraryType
 from virtool_workflow.analysis.read_prep.fastqc import fastqc
 from virtool_workflow.analysis.read_prep.skewer import (skewer,
                                                         trimming_min_length)
@@ -50,14 +55,16 @@ class Trimming(WorkflowFeature):
     async def __modify_environment__(self, environment: WorkflowEnvironment):
         environment.override("sample", self._sample_fixture)
 
-    async def _sample_fixture(self,
-                              sample_provider: SampleProvider,
-                              work_path: Path,
-                              run_subprocess: RunSubprocess,
-                              run_in_executor: FunctionExecutor,
-                              http: aiohttp.ClientSession,
-                              jobs_api_url: str,
-                              scope: FixtureScope):
+    async def _sample_fixture(
+            self,
+            sample_provider: SampleProvider,
+            work_path: Path,
+            run_subprocess: RunSubprocess,
+            run_in_executor: FunctionExecutor,
+            http: aiohttp.ClientSession,
+            jobs_api_url: str,
+            scope: FixtureScope
+    ):
         """Alternate sample fixture which performs trimming and caching of reads."""
         sample = await sample_provider.get()
         key = self._compute_cache_key(sample)
@@ -147,3 +154,35 @@ class Trimming(WorkflowFeature):
             cache.quality = quality
 
             return cache
+
+
+def trimming_parameters(
+        library_type: LibraryType,
+        trimming_min_length: int
+) -> Dict[str, Union[str, int]]:
+    """
+    Calculates trimming parameters based on the library type, and minimum allowed trim length.
+
+    :param library_type: The LibraryType (eg. srna)
+    :param trimming_min_length: The minimum length of a read before it is discarded.
+    :return: the trimming parameters
+    """
+    if library_type == LibraryType.amplicon:
+        return {
+            **TRIM_PARAMETERS,
+            "end_quality": 0,
+            "mean_quality": 0,
+            "min_length": trimming_min_length
+        }
+
+    if library_type == LibraryType.srna:
+        return {
+            **TRIM_PARAMETERS,
+            "min_length": 20,
+            "max_length": 22,
+        }
+
+    return {
+        **TRIM_PARAMETERS,
+        "min_length": trimming_min_length
+    }
